requirements.txt
autopep8==2.3.2
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-Session==0.8.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
pluggy==1.6.0
pycodestyle==2.14.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.1
pytest==8.4.2
pytest-mock==3.15.1
python-dotenv==1.1.1
requests==2.32.5
tqdm==4.67.1
urllib3==2.5.0
vercel_blob==0.4.2
Werkzeug==3.1.3
wonderwords==2.2.0
WTForms==3.2.1
